+++ flink-table/flink-table-planner/src/main/java/org/apache/flink/table/planner/plan/nodes/exec/ExecEdge.java
@@ -22,7 +22,6 @@ import org.apache.flink.api.dag.Transformation;
 import org.apache.flink.streaming.api.transformations.StreamExchangeMode;
 import org.apache.flink.table.api.TableException;
 import org.apache.flink.table.delegation.Planner;
-import org.apache.flink.table.planner.plan.nodes.exec.serde.ExecNodeGraphJsonPlanGenerator.JsonPlanEdge;
 import org.apache.flink.table.types.logical.LogicalType;
 
 import java.util.Arrays;
@@ -33,10 +32,10 @@ import static org.apache.flink.util.Preconditions.checkNotNull;
 /**
  * The representation of an edge connecting two {@link ExecNode}s.
  *
- * <p>The edge's json serialization/deserialization will be delegated to {@link JsonPlanEdge}, which
+ * <p>The edge's json serialization/deserialization will be delegated to {@code JsonPlanEdge}, which
  * only stores the {@link ExecNode}'s id instead of instance.
  *
- * <p>{@link JsonPlanEdge} should also be updated with this class if the fields are added/removed.
+ * <p>{@code JsonPlanEdge} should also be updated with this class if the fields are added/removed.
  */
 public class ExecEdge {
     /** The source node of this edge. */
